Refactor the people page, making it easier to edit.
When the primary purpose of the page is to list names and websites,
cross referencing the two makes it hard to navigate the code.

It may not look pretty to have the links and the names together,
but it is easier to edit than having a cross reference in the source code.

This is an example of anti-pattern. @kbarbary